+++ docs/conf.py
@@ -33,12 +33,27 @@ yml_ansible_roles = '../ansible/roles/'
 # in reStructuredText
 for element in os.listdir(rst_ansible_roles):
     if os.path.isdir(yml_ansible_roles + element + '/defaults'):
-        yaml2rst.convert_file(
-            yml_ansible_roles + element + '/defaults/main.yml',
-            rst_ansible_roles + element + '/defaults.rst',
-            strip_regex=r'\s*(:?\[{3}|\]{3})\d?$',
-            yaml_strip_regex=r'^\s{66,67}#\s\]{3}\d?$',
-        )
+        for path, subdirs, files in os.walk(yml_ansible_roles +
+                                            element + '/defaults'):
+            for filename in files:
+                if not filename.startswith('.'):
+                    defaults_file = os.path.join(path, filename)
+                    defaults_dir = os.path.dirname(defaults_file).lstrip('../')
+
+                    if not os.path.isdir(defaults_dir):
+                        try:
+                            os.makedirs(defaults_dir)
+                        except OSError:
+                            print ("Creation of the directory %s failed"
+                                   % defaults_dir)
+
+                    yaml2rst.convert_file(
+                        defaults_file,
+                        (os.path.splitext(defaults_file)[0]
+                            + '.rst').lstrip('../'),
+                        strip_regex=r'\s*(:?\[{3}|\]{3})\d?$',
+                        yaml_strip_regex=r'^\s{66,67}#\s\]{3}\d?$',
+                    )
 
 
 # Ignore warnings about non-local images